refactor generateText tests
