+++ packages/ai/src/telemetry/open-telemetry-integration.test.ts
@@ -1,3 +1,5 @@
+import { LanguageModelV4Usage } from '@ai-sdk/provider';
+import { tool } from '@ai-sdk/provider-utils';
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import {
   Attributes,
@@ -6,8 +8,17 @@ import {
   SpanStatusCode,
   Tracer,
 } from '@opentelemetry/api';
+import { z } from 'zod/v4';
+import { generateText } from '../generate-text/generate-text';
+import { stepCountIs } from '../generate-text/stop-condition';
+import { MockLanguageModelV4 } from '../test/mock-language-model-v4';
+import { MockTracer as IntegrationMockTracer } from '../test/mock-tracer';
 import { OpenTelemetryIntegration } from './open-telemetry-integration';
 import type { TelemetryIntegration } from './telemetry-integration';
+
+vi.mock('../version', () => ({
+  VERSION: '0.0.0-test',
+}));
 
 type MockSpan = Span & {
   name: string;
@@ -1198,5 +1209,540 @@ describe('OpenTelemetryIntegration', () => {
       expect(attrs['ai.telemetry.metadata.userId']).toBe('user-123');
       expect(attrs['ai.telemetry.metadata.sessionId']).toBe('sess-456');
     });
+  });
+});
+
+const integrationTestUsage: LanguageModelV4Usage = {
+  inputTokens: {
+    total: 3,
+    noCache: 3,
+    cacheRead: undefined,
+    cacheWrite: undefined,
+  },
+  outputTokens: {
+    total: 10,
+    text: 10,
+    reasoning: undefined,
+  },
+};
+
+const integrationDummyResponseValues = {
+  finishReason: { unified: 'stop', raw: 'stop' } as const,
+  usage: integrationTestUsage,
+  warnings: [],
+};
+
+const integrationModelWithReasoning = new MockLanguageModelV4({
+  doGenerate: {
+    ...integrationDummyResponseValues,
+    content: [
+      {
+        type: 'reasoning',
+        text: 'I will open the conversation with witty banter.',
+        providerMetadata: {
+          testProvider: {
+            signature: 'signature',
+          },
+        },
+      },
+      {
+        type: 'reasoning',
+        text: '',
+        providerMetadata: {
+          testProvider: {
+            redactedData: 'redacted-reasoning-data',
+          },
+        },
+      },
+      { type: 'text', text: 'Hello, world!' },
+    ],
+  },
+});
+
+describe('OpenTelemetryIntegration integration with generateText', () => {
+  let tracer: IntegrationMockTracer;
+
+  beforeEach(() => {
+    tracer = new IntegrationMockTracer();
+  });
+
+  it('should not record any telemetry data when not explicitly enabled', async () => {
+    await generateText({
+      model: new MockLanguageModelV4({
+        doGenerate: async ({}) => ({
+          ...integrationDummyResponseValues,
+          content: [{ type: 'text', text: 'Hello, world!' }],
+        }),
+      }),
+      prompt: 'prompt',
+      experimental_telemetry: {
+        integrations: new OpenTelemetryIntegration({ tracer }),
+      },
+    });
+
+    expect(tracer.jsonSpans).toMatchSnapshot();
+  });
+
+  it('should record telemetry data when enabled', async () => {
+    await generateText({
+      model: new MockLanguageModelV4({
+        doGenerate: async ({}) => ({
+          ...integrationDummyResponseValues,
+          content: [{ type: 'text', text: 'Hello, world!' }],
+          response: {
+            id: 'test-id-from-model',
+            timestamp: new Date(10000),
+            modelId: 'test-response-model-id',
+          },
+          providerMetadata: {
+            testProvider: {
+              testKey: 'testValue',
+            },
+          },
+        }),
+      }),
+      prompt: 'prompt',
+      topK: 0.1,
+      topP: 0.2,
+      frequencyPenalty: 0.3,
+      presencePenalty: 0.4,
+      temperature: 0.5,
+      stopSequences: ['stop'],
+      headers: {
+        header1: 'value1',
+        header2: 'value2',
+      },
+      experimental_telemetry: {
+        isEnabled: true,
+        functionId: 'test-function-id',
+        metadata: {
+          test1: 'value1',
+          test2: false,
+        },
+        integrations: new OpenTelemetryIntegration({ tracer }),
+      },
+    });
+
+    expect(tracer.jsonSpans).toMatchSnapshot();
+  });
+
+  it('should record successful tool call', async () => {
+    await generateText({
+      model: new MockLanguageModelV4({
+        doGenerate: async ({}) => ({
+          ...integrationDummyResponseValues,
+          content: [
+            {
+              type: 'tool-call',
+              toolCallType: 'function',
+              toolCallId: 'call-1',
+              toolName: 'tool1',
+              input: `{ "value": "value" }`,
+            },
+          ],
+          response: {
+            id: 'test-id',
+            timestamp: new Date(0),
+            modelId: 'mock-model-id',
+          },
+        }),
+      }),
+      tools: {
+        tool1: {
+          inputSchema: z.object({ value: z.string() }),
+          execute: async () => 'result1',
+        },
+      },
+      prompt: 'test-input',
+      experimental_telemetry: {
+        isEnabled: true,
+        integrations: new OpenTelemetryIntegration({ tracer }),
+      },
+      _internal: {
+        generateId: () => 'test-id',
+        generateCallId: () => 'test-telemetry-call-id',
+      },
+    });
+
+    expect(tracer.jsonSpans).toMatchInlineSnapshot(`
+      [
+        {
+          "attributes": {
+            "ai.model.id": "mock-model-id",
+            "ai.model.provider": "mock-provider",
+            "ai.operationId": "ai.generateText",
+            "ai.prompt": "{"prompt":"test-input"}",
+            "ai.request.headers.user-agent": "ai/0.0.0-test",
+            "ai.response.finishReason": "stop",
+            "ai.response.text": "",
+            "ai.response.toolCalls": "[{"toolCallId":"call-1","toolName":"tool1","input":{"value":"value"}}]",
+            "ai.settings.maxRetries": 2,
+            "ai.usage.inputTokenDetails.noCacheTokens": 3,
+            "ai.usage.inputTokens": 3,
+            "ai.usage.outputTokenDetails.textTokens": 10,
+            "ai.usage.outputTokens": 10,
+            "ai.usage.totalTokens": 13,
+            "operation.name": "ai.generateText",
+          },
+          "events": [],
+          "name": "ai.generateText",
+        },
+        {
+          "attributes": {
+            "ai.model.id": "mock-model-id",
+            "ai.model.provider": "mock-provider",
+            "ai.operationId": "ai.generateText.doGenerate",
+            "ai.prompt.messages": "[{"role":"user","content":[{"type":"text","text":"test-input"}]}]",
+            "ai.prompt.toolChoice": "{"type":"auto"}",
+            "ai.prompt.tools": [
+              "{"type":"function","name":"tool1","inputSchema":{"$schema":"http://json-schema.org/draft-07/schema#","type":"object","properties":{"value":{"type":"string"}},"required":["value"],"additionalProperties":false}}",
+            ],
+            "ai.request.headers.user-agent": "ai/0.0.0-test",
+            "ai.response.finishReason": "stop",
+            "ai.response.id": "test-id",
+            "ai.response.model": "mock-model-id",
+            "ai.response.text": "",
+            "ai.response.timestamp": "1970-01-01T00:00:00.000Z",
+            "ai.response.toolCalls": "[{"toolCallId":"call-1","toolName":"tool1","input":{"value":"value"}}]",
+            "ai.settings.maxRetries": 2,
+            "ai.usage.inputTokenDetails.noCacheTokens": 3,
+            "ai.usage.inputTokens": 3,
+            "ai.usage.outputTokenDetails.textTokens": 10,
+            "ai.usage.outputTokens": 10,
+            "ai.usage.totalTokens": 13,
+            "gen_ai.request.model": "mock-model-id",
+            "gen_ai.response.finish_reasons": [
+              "stop",
+            ],
+            "gen_ai.response.id": "test-id",
+            "gen_ai.response.model": "mock-model-id",
+            "gen_ai.system": "mock-provider",
+            "gen_ai.usage.input_tokens": 3,
+            "gen_ai.usage.output_tokens": 10,
+            "operation.name": "ai.generateText.doGenerate",
+          },
+          "events": [],
+          "name": "ai.generateText.doGenerate",
+        },
+        {
+          "attributes": {
+            "ai.operationId": "ai.toolCall",
+            "ai.toolCall.args": "{"value":"value"}",
+            "ai.toolCall.id": "call-1",
+            "ai.toolCall.name": "tool1",
+            "ai.toolCall.result": ""result1"",
+            "operation.name": "ai.toolCall",
+          },
+          "events": [],
+          "name": "ai.toolCall",
+        },
+      ]
+    `);
+  });
+
+  it('should record error on tool call', async () => {
+    await generateText({
+      model: new MockLanguageModelV4({
+        doGenerate: async ({}) => ({
+          ...integrationDummyResponseValues,
+          content: [
+            {
+              type: 'tool-call',
+              toolCallType: 'function',
+              toolCallId: 'call-1',
+              toolName: 'tool1',
+              input: `{ "value": "value" }`,
+            },
+          ],
+        }),
+      }),
+      tools: {
+        tool1: {
+          inputSchema: z.object({ value: z.string() }),
+          execute: async () => {
+            throw new Error('Tool execution failed');
+          },
+        },
+      },
+      prompt: 'test-input',
+      experimental_telemetry: {
+        isEnabled: true,
+        integrations: new OpenTelemetryIntegration({ tracer }),
+      },
+      _internal: {
+        generateId: () => 'test-id',
+        generateCallId: () => 'test-telemetry-call-id',
+      },
+    });
+
+    expect(tracer.jsonSpans).toHaveLength(3);
+
+    expect(tracer.jsonSpans[0].name).toBe('ai.generateText');
+    expect(tracer.jsonSpans[1].name).toBe('ai.generateText.doGenerate');
+    expect(tracer.jsonSpans[2].name).toBe('ai.toolCall');
+
+    const toolCallSpan = tracer.jsonSpans[2];
+    expect(toolCallSpan.status).toEqual({
+      code: 2,
+      message: 'Tool execution failed',
+    });
+
+    expect(toolCallSpan.events).toHaveLength(1);
+    const exceptionEvent = toolCallSpan.events[0];
+    expect(exceptionEvent.name).toBe('exception');
+    expect(exceptionEvent.attributes).toMatchObject({
+      'exception.message': 'Tool execution failed',
+      'exception.name': 'Error',
+    });
+    expect(exceptionEvent.attributes?.['exception.stack']).toContain(
+      'Tool execution failed',
+    );
+    expect(exceptionEvent.time).toEqual([0, 0]);
+  });
+
+  it('should not record telemetry inputs / outputs when disabled', async () => {
+    await generateText({
+      model: new MockLanguageModelV4({
+        doGenerate: async ({}) => ({
+          ...integrationDummyResponseValues,
+          content: [
+            {
+              type: 'tool-call',
+              toolCallType: 'function',
+              toolCallId: 'call-1',
+              toolName: 'tool1',
+              input: `{ "value": "value" }`,
+            },
+          ],
+          response: {
+            id: 'test-id',
+            timestamp: new Date(0),
+            modelId: 'mock-model-id',
+          },
+        }),
+      }),
+      tools: {
+        tool1: {
+          inputSchema: z.object({ value: z.string() }),
+          execute: async () => 'result1',
+        },
+      },
+      prompt: 'test-input',
+      experimental_telemetry: {
+        isEnabled: true,
+        recordInputs: false,
+        recordOutputs: false,
+        integrations: new OpenTelemetryIntegration({ tracer }),
+      },
+      _internal: {
+        generateId: () => 'test-id',
+        generateCallId: () => 'test-telemetry-call-id',
+      },
+    });
+
+    expect(tracer.jsonSpans).toMatchSnapshot();
+  });
+
+  it('should record reasoning in telemetry when present', async () => {
+    await generateText({
+      model: integrationModelWithReasoning,
+      prompt: 'test-input',
+      experimental_telemetry: {
+        isEnabled: true,
+        integrations: new OpenTelemetryIntegration({ tracer }),
+      },
+    });
+
+    const rootSpan = tracer.jsonSpans.find(
+      span => span.name === 'ai.generateText',
+    );
+    const doGenerateSpan = tracer.jsonSpans.find(
+      span => span.name === 'ai.generateText.doGenerate',
+    );
+
+    expect(rootSpan?.attributes['ai.response.reasoning']).toBe(
+      'I will open the conversation with witty banter.\n',
+    );
+    expect(doGenerateSpan?.attributes['ai.response.reasoning']).toBe(
+      'I will open the conversation with witty banter.\n',
+    );
+  });
+
+  it('should record total usage across steps in root span', async () => {
+    let responseCount = 0;
+    await generateText({
+      model: new MockLanguageModelV4({
+        doGenerate: async () => {
+          switch (responseCount++) {
+            case 0:
+              return {
+                finishReason: {
+                  unified: 'tool-calls' as const,
+                  raw: undefined,
+                },
+                usage: {
+                  inputTokens: {
+                    total: 5,
+                    noCache: 5,
+                    cacheRead: undefined,
+                    cacheWrite: undefined,
+                  },
+                  outputTokens: {
+                    total: 20,
+                    text: 20,
+                    reasoning: undefined,
+                  },
+                },
+                warnings: [],
+                content: [
+                  {
+                    type: 'tool-call' as const,
+                    toolCallType: 'function' as const,
+                    toolCallId: 'call-1',
+                    toolName: 'tool1',
+                    input: '{ "value": "value" }',
+                  },
+                ],
+              };
+            case 1:
+            default:
+              return {
+                finishReason: {
+                  unified: 'stop' as const,
+                  raw: 'stop',
+                },
+                usage: {
+                  inputTokens: {
+                    total: 10,
+                    noCache: 10,
+                    cacheRead: undefined,
+                    cacheWrite: undefined,
+                  },
+                  outputTokens: {
+                    total: 15,
+                    text: 15,
+                    reasoning: undefined,
+                  },
+                },
+                warnings: [],
+                content: [{ type: 'text' as const, text: 'Final answer.' }],
+              };
+          }
+        },
+      }),
+      tools: {
+        tool1: {
+          inputSchema: z.object({ value: z.string() }),
+          execute: async () => 'result1',
+        },
+      },
+      stopWhen: stepCountIs(2),
+      prompt: 'test-input',
+      experimental_telemetry: {
+        isEnabled: true,
+        integrations: new OpenTelemetryIntegration({ tracer }),
+      },
+    });
+
+    const rootSpan = tracer.jsonSpans.find(
+      span => span.name === 'ai.generateText',
+    );
+
+    expect(rootSpan?.attributes['ai.usage.inputTokens']).toBe(15);
+    expect(rootSpan?.attributes['ai.usage.outputTokens']).toBe(35);
+    expect(rootSpan?.attributes['ai.usage.totalTokens']).toBe(50);
+  });
+
+  it('should execute subagent generateText inside executeToolCall context', async () => {
+    let activeContext: string | undefined;
+    let capturedContext: string | undefined;
+    const otelIntegration = new OpenTelemetryIntegration({ tracer });
+
+    await generateText({
+      model: new MockLanguageModelV4({
+        doGenerate: async () => ({
+          ...integrationDummyResponseValues,
+          content: [
+            {
+              type: 'tool-call',
+              toolCallId: 'call-1',
+              toolName: 'researchTool',
+              input: '{ "city": "Tokyo" }',
+            },
+          ],
+        }),
+      }),
+      tools: {
+        researchTool: tool({
+          inputSchema: z.object({ city: z.string() }),
+          execute: async ({ city }) => {
+            capturedContext = activeContext;
+
+            const subResult = await generateText({
+              model: new MockLanguageModelV4({
+                doGenerate: async () => ({
+                  ...integrationDummyResponseValues,
+                  content: [
+                    { type: 'text', text: `Weather in ${city}: sunny` },
+                  ],
+                }),
+              }),
+              prompt: `Weather for ${city}`,
+              experimental_telemetry: {
+                isEnabled: true,
+                functionId: `sub-agent-${city.toLowerCase()}`,
+                integrations: otelIntegration,
+              },
+            });
+            return subResult.text;
+          },
+        }),
+      },
+      prompt: 'test-input',
+      experimental_telemetry: {
+        isEnabled: true,
+        functionId: 'weather-agent',
+        integrations: [
+          otelIntegration,
+          {
+            executeTool: async ({ callId, toolCallId, execute }) => {
+              activeContext = `${callId}:${toolCallId}`;
+              try {
+                return await execute();
+              } finally {
+                activeContext = undefined;
+              }
+            },
+          },
+        ],
+      },
+      _internal: {
+        generateId: () => 'outer-test-id',
+        generateCallId: () => 'outer-test-call-id',
+      },
+    });
+
+    expect(capturedContext).toBe('outer-test-call-id:call-1');
+
+    expect(tracer.spans.map(s => s.name)).toEqual([
+      'ai.generateText',
+      'ai.generateText.doGenerate',
+      'ai.toolCall',
+      'ai.generateText',
+      'ai.generateText.doGenerate',
+    ]);
+
+    const toolCallSpan = tracer.spans[2];
+    expect(toolCallSpan.attributes['ai.toolCall.name']).toBe('researchTool');
+
+    const innerRootSpan = tracer.spans[3];
+    expect(innerRootSpan.attributes['ai.telemetry.functionId']).toBe(
+      'sub-agent-tokyo',
+    );
+
+    const innerStepSpan = tracer.spans[4];
+    expect(innerStepSpan.attributes['ai.response.text']).toBe(
+      'Weather in Tokyo: sunny',
+    );
   });
 });
